metabase: Move all utils out of management commands folder
