+++ itou/metabase/sql/002_missions_ai_ehpad.sql
