+++ itou/metabase/sql/004_fluxIAE_Structure_v2.sql
