+++ itou/metabase/sql/018_nombre_candidatures_par_structure.sql
