+++ itou/metabase/sql/023_candidats_sans_solution.sql
