+++ itou/metabase/tables/siaes.py
@@ -2,7 +2,7 @@ from django.utils import timezone
 
 from itou.job_applications.enums import SenderKind
 from itou.job_applications.models import JobApplicationWorkflow
-from itou.metabase.management.commands._utils import (
+from itou.metabase.tables.utils import (
     MetabaseTable,
     get_address_columns,
     get_choice,
